hoàn thành

+++ src/service/impl/CustomerService.java
@@ -2,8 +2,13 @@ package service.impl;
 
 import dao.ICustomerDao;
 import dao.impl.CustomerDao;
+import dao.impl.IvoiceDao;
 import model.Customer;
+import model.Invoice;
+import presentation.UserView;
 import service.ICustomerService;
+
+import java.util.List;
 
 public class CustomerService implements ICustomerService {
     ICustomerDao customerDao = new CustomerDao();
@@ -24,6 +29,15 @@ public class CustomerService implements ICustomerService {
 
     @Override
     public void deleteCustomer(int id) {
-        customerDao.deleteCustomer(id);
+
+       List<Invoice> a =IvoiceDao.findInvoicebyidProdusct(id);
+       if(a!=null){
+           System.out.println(UserView.RED+"Khác hàng này đã mua hàng khổng thể xóa"+UserView.RESET);
+       }
+       else {
+           customerDao.deleteCustomer(id);
+           System.out.println("=> Đã xóa khách hàng thành công!");
+       }
+
     }
 }
